WIP: Move strategy to dependency
There are still strategy related tests here now, but they will be removed.
We are only leaving in this repo code related to the framework itself.

The idea is that we will import the strategies. The code here will
only do the plumbing to run the strategy using the Producers/Consumers of
different exchanges.